Unbelivable

+++ Solution/src/Hilbert.hs
@@ -3,6 +3,7 @@ module Hilbert where
 import Data.Bits
 import Data.List
 import Settings
+import Data.Int
 
 grid :: [(Int, Int)]
 grid = nub $ map (scale . d2xy 32) [0..32 * 32 - 1]
@@ -14,7 +15,7 @@ gridDown :: [(Int, Int)]
 gridDown = map (shiftHilbert (0, 3)) grid
 
 shiftHilbert :: (Int, Int) -> (Int, Int) -> (Int, Int)
-shiftHilbert (dx, dy) (x, y) = ((x + dx) `mod` gridWidth, (y + dy) `mod` gridHeight)
+shiftHilbert (dx, dy) (x, y) = ((x + dx) `mod` fromIntegral gridWidth, (y + dy) `mod` fromIntegral gridHeight)
 
 scale :: (Int, Int) -> (Int, Int)
 scale (x, y) = (floor (fromIntegral x / 32 * 24), floor (fromIntegral y / 32 * 24))
